done MRKT PLACE and PUBLISH
I have restructured the smart contract, by making a big struct called Coin with takes all the variables present in Offer and Bid wich existed before. mrktplace and publish are now working and render photos, still need to work on button for using the function such as Bid, Sell etc etc. Still need to work on the definition of the smart contract.